Fix narrowing cast of time_t in resumption data (#3946)

+++ src/components/application_manager/include/application_manager/resumption/resumption_data.h
@@ -157,7 +157,7 @@ class ResumptionData {
    * @brief Get the last ignition off time from LastState
    * @return the last ignition off time from LastState
    */
-  virtual uint32_t GetIgnOffTime() const = 0;
+  virtual int64_t GetIgnOffTime() const = 0;
 
   /**
    * @brief Checks if saved data have application
